utilities/requirement.txt
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
Babel==2.17.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
csvw==3.5.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dlinfo==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fonttools==4.58.1
fsspec==2025.5.1
hf-xet==1.1.3
huggingface-hub==0.32.4
HyperPyYAML==1.2.2
idna==3.10
importlib-metadata==8.7.0
ipykernel==6.29.5
ipython==9.3.0
ipython-pygments-lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.8.1
kiwisolver==1.4.8
language-tags==1.2.0
lazy-loader==0.4
librosa==0.11.0
lightning-utilities==0.14.3
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.0
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
phonemizer==3.3.0
pickleshare==0.7.5
Pillow==11.2.1
pip==25.1.1
platformdirs==4.3.8
pooch==1.8.2
prompt-toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
py-espeak-ng==0.1.8
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.25.1
ruamel.yaml==0.18.12
ruamel.yaml.clib==0.2.12
s3prl==0.4.18
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
segments==2.3.0
sentencepiece==0.2.0
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.3
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
torchaudio==2.7.1
torchmetrics==1.7.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
triton==3.3.1
typing-extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
wcwidth==0.2.13
wheel==0.45.1
zipp==3.22.0
